more work on asset manager
